+++ network-api/networkapi/minisites/templates/minisites/blocks/iframe_block.html
@@ -1,7 +1,7 @@
 {% load wagtailcore_tags %}
 
 <iframe
-    width="{{ block.value.width }}"
+    width="100%"
     height="{{ block.value.height }}"
     src="{{ block.value.url }}"
     frameborder="0"
